% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{implied_handler}
\alias{implied_handler}
\title{Handles the calculations in implied_probabilities}
\usage{
implied_handler(odds, method = "method")
}
\arguments{
\item{odds}{Odds in a preprocessed format.}

\item{method}{The method, already preprocessed from `implied_probabilities`.}
}
\value{
An `out` list for use `implied_probablities`.
}
\description{
Handles the calculations in implied_probabilities
}
