% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{or_func}
\alias{or_func}
\title{Calculate the probabilities using the odds ratio method, for a given value
of the odds ratio cc.}
\usage{
or_func(cc, io)
}
\arguments{
\item{io}{Inverted odds.}

\item{zz}{?}
}
\value{
Probabilities using the odds ratio method.
}
\description{
Calculate the probabilities using the odds ratio method, for a given value
of the odds ratio cc.
}
\keyword{internal}
