% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_probabilities.R
\name{or_solvefor}
\alias{or_solvefor}
\title{The condition that the sum of the probabilites must sum to 1. This function
calulates the true probability, given bookmaker probabilites xx, and the
odds ratio cc.}
\usage{
or_solvefor(cc, io)
}
\arguments{
\item{cc}{?}

\item{io}{Inverted odds.}
}
\value{
True probabilities.
}
\description{
The condition that the sum of the probabilites must sum to 1. This function
calulates the true probability, given bookmaker probabilites xx, and the
odds ratio cc.
}
\keyword{internal}
