% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_probabilities.R
\name{pwr_func}
\alias{pwr_func}
\title{Power function}
\usage{
pwr_func(nn, io)
}
\arguments{
\item{nn}{Numeric vector.}

\item{io}{?.}
}
\value{
Numeric.
}
\description{
Power function
}
\keyword{internal}
