% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{pwr_solvefor}
\alias{pwr_solvefor}
\title{The condition that the sum of the probabilites must sum to 1. This function
calulates the true probability, given bookmaker probabilites xx, and the
inverse exponent.}
\usage{
pwr_solvefor(nn, io)
}
\arguments{
\item{nn}{Numeric vector.}

\item{io}{?.}
}
\value{
True probabilities.
}
\description{
The condition that the sum of the probabilites must sum to 1. This function
calulates the true probability, given bookmaker probabilites xx, and the
inverse exponent.
}
\keyword{internal}
