% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{shin_func}
\alias{shin_func}
\title{Calculate the probabilities using Shin's formula, for a given value of z.}
\usage{
shin_func(zz, io)
}
\arguments{
\item{zz}{?}

\item{io}{Inverted odds.}
}
\value{
Probabilities using Shin's formula
}
\description{
Calculate the probabilities using Shin's formula, for a given value of z.
}
\keyword{internal}
