% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{shin_solvefor}
\alias{shin_solvefor}
\title{The condition that the sum of the probabilites must sum to 1. Used with
uniroot.}
\usage{
shin_solvefor(zz, io)
}
\arguments{
\item{zz}{?}

\item{io}{Inverted odds.}
}
\value{
`0` if the probabilities sum to 1, `1` otherwise.
}
\description{
The condition that the sum of the probabilites must sum to 1. Used with
uniroot.
}
\keyword{internal}
