% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_probabilities.R
\name{implied_probabilities}
\alias{implied_probabilities}
\title{Implied probabilities from bookmaker odds.}
\usage{
implied_probabilities(odds, method = "basic", normalize = TRUE)
}
\arguments{
\item{odds}{A matrix or numeric of bookmaker odds. The odds must be in the decimal format.}

\item{method}{A string giving the method to use. Valid methods are 'basic', 'shin',
'wpo', 'or' and 'power'.}

\item{normalize}{Logical. Some of the methods will give small rounding errors. If TRUE (default)
a final normalization is applied to make absoultely sure the
probabilities sum to 1.

The method 'basic' is the simplest method, and computes the implied probabilities by
dividing the inverted odds by the sum of the inverted odds.

The method 'shin' uses the method by Shin (1991). This model assumes that there is a fraction of
insider trading, and that the bookmakers tries to maximize their profits. In addition to providing
implied probabiltiies, the method also gives an estimate of the proportion if inside trade. The method
implemented here is based on the description in Štrumbelj (2014).

The methods 'wpo', 'or' and 'power' are form the Wisdom of the Crowds document (the updated version) by
Joseph Buchdahl. The method 'or' is origianlly by Cheung (2015), and the method 'power' is there referred
to as the logarithmic method.}
}
\value{
A named list. The first component is named 'probabilities' and contain a matrix of
implied probabilities. The second in the bookmaker margins (aka the overround). The third
depends on the method used to compute the probabilities:
 * zvalues (method = 'shin'): The estimated amount of insider trade.
 * specific_margins (method = 'wpo'): Matrix of the margins applied to each outcome.
 * odds_ratios (method = 'or'): Numeric with the odds ratio that is used to convert true
 probabilities to bookamker probabilties.
 * exponents (method = 'power'): The (inverse) exponents that is used to convert true
 probabilities to bookamker probabilties.

 @section References
 * Hyun Song Shin (1991) Optimal betting odds against insider traders.
 Erik Štrumbelj (2014) On determining probability forecasts from betting odds.
 * Joseph Buchdahl - USING THE WISDOM OF THE CROWD TO FIND VALUE IN A FOOTBALL MATCH BETTING MARKET (http://www.football-data.co.uk/wisdom_of_crowd_bets)
 * Keith Cheung (2015) Fixed-odds betting and traditional odds (www.sportstradingnetwork.com/article/fixed-odds-betting-traditional-odds/)
}
\description{
This function calculate the implied probabilties from bookmaker odds in decimal format, using five
different methods to account for overround in the odds.
}
